+++ src/test-block.ts
@@ -6,12 +6,12 @@ export class TestBlock {
   constructor(protected api: ApiWrapper, protected props: ITest & IWithName) {
   }
 
-  public run() {
-    this.api.it(this.props, this.props.test);
-  }
-
   public getProps() {
     return this.props;
+  }
+
+  public run(context: any) {
+    this.api.it(this.props, this.props.test.bind(context));
   }
 
 }